Split up code logically
