+++ src/requirements.txt
@@ -3,3 +3,4 @@ requests==2.28.2
 PyYAML==6.0
 SQLAlchemy==2.0.9
 alembic==1.10.3
+discord==2.2.2
